+++ packages/tools/src/stateManagement/segmentation/segmentIndex.ts
@@ -22,27 +22,6 @@ function setActiveSegmentIndex(
   }
 }
 
-function setPreviewSegmentIndex(segmentationId, segmentIndex: number) {
-  const segmentation = getSegmentation(segmentationId);
-
-  if (segmentation?.previewSegmentIndex !== segmentIndex) {
-    segmentation.previewSegmentIndex = segmentIndex;
-
-    triggerSegmentationModified(segmentationId);
-  }
-}
-
-function getPreviewSegmentIndex(
-  segmentationId,
-  segmentationIndex?: number
-): number | undefined {
-  const segmentation = getSegmentation(segmentationId);
-
-  if (segmentation) {
-    return segmentation.previewSegmentIndex;
-  }
-}
-
 /**
  * Get the active segment index for a segmentation in the global state
  * @param segmentationId - The id of the segmentation to get the active segment index from.
@@ -56,9 +35,4 @@ function getActiveSegmentIndex(segmentationId: string): number | undefined {
   }
 }
 
-export {
-  getActiveSegmentIndex,
-  setActiveSegmentIndex,
-  getPreviewSegmentIndex,
-  setPreviewSegmentIndex,
-};
+export { getActiveSegmentIndex, setActiveSegmentIndex };