+++ packages/tools/src/tools/segmentation/strategies/utils/initializePreview.ts
@@ -1,14 +1,22 @@
 import type { InitializedOperationData } from '../BrushStrategy';
 import { triggerSegmentationDataModified } from '../../../../stateManagement/segmentation/triggerSegmentationEvents';
-import { segmentIndex as segmentIndexController } from '../../../../stateManagement/segmentation';
+import {
+  segmentIndex as segmentIndexController,
+  config as segmentationConfig,
+} from '../../../../stateManagement/segmentation';
 /**
  * Sets up a preview to use an alternate set of colours.  First fills the
  * preview segment index with the final one for all pixels, then resets
  * the preview colours.
+ * This is only activated when the preview segment index is defined, eihter
+ * from the initial state or from the global state.
  */
 export default {
   preview: function (enabled, operationData: InitializedOperationData) {
-    operationData.previewSegmentIndex ??= 4;
+    const { previewSegmentIndex } = operationData;
+    if (!previewSegmentIndex) {
+      return;
+    }
     this.initDown?.(enabled, operationData);
     const preview = this.fill(enabled, operationData);
     operationData.preview = preview;
@@ -17,16 +25,30 @@ export default {
   },
 
   createInitialized: (enabled, operationData: InitializedOperationData) => {
-    const { segmentationId } = operationData;
-    const previewSegmentIndex =
-      segmentIndexController.getPreviewSegmentIndex(segmentationId);
-
-    if (
-      previewSegmentIndex !== undefined &&
-      operationData.previewSegmentIndex === undefined
-    ) {
-      operationData.previewSegmentIndex = previewSegmentIndex;
+    if (operationData.previewSegmentIndex === undefined) {
+      return;
     }
+    const {
+      toolGroupId,
+      segmentIndex,
+      segmentationRepresentationUID,
+      previewSegmentIndex,
+      previewColors,
+    } = operationData;
+
+    const configColor = previewColors?.[segmentIndex];
+    const segmentColor = segmentationConfig.color.getColorForSegmentIndex(
+      toolGroupId,
+      segmentationRepresentationUID,
+      segmentIndex
+    );
+    const previewColor = configColor || segmentColor.map((it) => it * 0.9);
+    segmentationConfig.color.setColorForSegmentIndex(
+      toolGroupId,
+      segmentationRepresentationUID,
+      previewSegmentIndex,
+      previewColor as [number, number, number, number]
+    );
   },
 
   acceptPreview: (enabledElement, operationData: InitializedOperationData) => {