+++ packages/tools/src/types/ClosestControlPoint.ts
@@ -1,0 +1,6 @@
+import type { ClosestPoint } from './ClosestPoint';
+
+export type ClosestControlPoint = ClosestPoint & {
+  /** Control point index */
+  index: number;
+};