+++ packages/tools/src/types/LabelmapToolOperationData.ts
@@ -1,3 +1,5 @@
+import type { Types } from '@cornerstonejs/core';
+
 import {
   LabelmapSegmentationDataStack,
   LabelmapSegmentationDataVolume,
@@ -6,12 +8,29 @@ import {
 type LabelmapToolOperationData = {
   segmentationId: string;
   segmentIndex: number;
+  /**
+   * The segment index to use for a preview
+   */
+  previewSegmentIndex?: number;
+  /**
+   * The colours to use for previewing
+   */
+  previewColors?: Record<number, [number, number, number, number]>;
   segmentsLocked: number[];
   viewPlaneNormal: number[];
   viewUp: number[];
   strategySpecificConfiguration: any;
-  constraintFn: (pointIJK: number) => boolean;
+  // constraintFn: (pointIJK: number) => boolean;
   segmentationRepresentationUID: string;
+  points: Types.Point3[];
+  representationUID: string;
+  /**
+   * preview is used for sharing preview data between views/interactions with
+   * a tool, and needs to be maintained by the tool side in order to be able
+   * to accept/reject/update the preview information.
+   */
+  preview: any;
+  toolGroupId: string;
 };
 
 type LabelmapToolOperationDataStack = LabelmapToolOperationData &