+++ packages/tools/src/utilities/math/aabb/index.ts
@@ -1,0 +1,2 @@
+export { default as distanceToPoint } from './distanceToPoint';
+export { default as distanceToPointSquared } from './distanceToPointSquared';